Make channel name restrictions work correctly
The problem is that `msg.envelope.room` is not the name of a channel, but
a channel ID. We were checking whether the channel ID was in a list of
channel names, which is obviously never going to work.

I copied over the `getRoomNameFromMessage` method from our PagerDuty
script, which is proven to work.